requirements.txt
requests
flask
flask-babel
flask-login
flask-sqlalchemy
flask-assets
flask-wtf

sqlalchemy
sqlalchemy-utils

loguru

python-dotenv

isort
black

Markdown
Pygments

pylint
dlint
flake8

github3.py
